djangobackend/requirements.txt
asgiref==3.5.2
astroid==2.13.3
autopep8==2.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
filelock==3.8.0
idna==3.4
inflection==0.5.1
isort==5.11.4
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.3
pycodestyle==2.10.0
pylint==2.15.10
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
virtualenvwrapper-win==1.2.7
wrapt==1.14.1
